model 2/weather_restoration/requirements.txt
torch
torchvision
torchaudio
opencv-python
matplotlib
tqdm
Pillow
torchmetrics
